Conflicts:
	service/src/main/java/pl/edu/agh/iisg/to/to2project/service/CategoryService.java
	service/src/main/java/pl/edu/agh/iisg/to/to2project/service/ExternalTransactionService.java
	service/src/main/java/pl/edu/agh/iisg/to/to2project/service/impl/CategoryServiceImpl.java
	service/src/main/java/pl/edu/agh/iisg/to/to2project/service/impl/ExternalTransactionServiceImpl.java

+++ app/src/main/java/pl/edu/agh/iisg/to/to2project/app/expenses/transactions/controller/DeleteTransactionPopupController.java
@@ -3,6 +3,7 @@ package pl.edu.agh.iisg.to.to2project.app.expenses.transactions.controller;
 import com.google.common.base.Preconditions;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.text.Text;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Controller;
@@ -14,6 +15,7 @@ import pl.edu.agh.iisg.to.to2project.service.InternalTransactionService;
 
 /**
  * @author Bartłomiej Grochal
+ * @author Wojciech Pachuta
  */
 @Controller
 @Scope("prototype")
@@ -28,22 +30,37 @@ public class DeleteTransactionPopupController extends PopupController {
     @Autowired
     private TransactionsController transactionsController;
 
+    @FXML
+    public Text errorLabel;
+
     private InternalTransaction internalTransaction;
+
     private ExternalTransaction externalTransaction;
 
     @FXML
     @Override
     protected void handleOKButtonClick(ActionEvent actionEvent) {
+        boolean removed = false;
         if(externalTransaction != null) {
-            externalTransactionService.remove(externalTransaction);
+            if(externalTransactionService.canDelete(externalTransaction)) {
+                externalTransactionService.remove(externalTransaction);
+                removed = true;
+            }
         }
         else if (internalTransaction != null) {
-            internalTransactionService.remove(internalTransaction);
+            if(internalTransactionService.canDelete(internalTransaction)) {
+                internalTransactionService.remove(internalTransaction);
+                removed = true;
+            }
         }
 
-        transactionsController.refreshContent();
-
-        closeDialog();
+        if(removed) {
+            transactionsController.refreshContent();
+            closeDialog();
+        }
+        else{
+            errorLabel.setText("Sorry, you can't delete this transaction.");
+        }
     }
 
     public void deleteTransaction(InternalTransaction transaction) {
